Phase 4: apply every review-code finding, and close the parity gap it found
The critical one: item 7 computed the expected content origin in floating point
while step 2 pads with integer division, so any asset whose content dimension
had the opposite parity to its canvas was measured against an x.5 origin no
image can sit on. Item 7 was unsatisfiable for roughly half the 974 rows, on
assets that were correct.

The review fixed it but left it untested — every fixture used 32x24 on a 64 px
canvas, all even, so the bug was invisible. Added a Theory over both pivots at
31x25, which cannot halve; reverting the fix fails both parametrisations.

Also: a per-pixel LINQ Min in item 5 and a per-group full-frame mask allocation
in step 4, both on M8-10's 942-asset path; an unguarded PixelMask setter that
would wrap a negative x onto the previous row; and the SkiaSharp ownership
convention, which was consistent but undocumented.

Numeric-literal sweep over the whole project: 66 classified, 0 invented.

+++ tools/AssetPipeline/Qa/Checks/CanvasAndPivotCheck.cs
@@ -118,9 +118,19 @@ public sealed class CanvasAndPivotCheck : IQaCheck
     /// distance between where its top-left corner is and where it would be.
     /// </summary>
     /// <remarks>
+    /// <para>
     /// 🔒 Measured against the content, not against a declaration. §C says the pivot is "declared in
     /// the atlas metadata", and a declaration that disagrees with where the subject actually sits is
     /// the failure this item exists to catch.
+    /// </para>
+    /// <para>
+    /// 🔒 <b>The centring tie is broken the same way <see cref="TrimToCanvasStep"/> breaks it</b> —
+    /// integer division, so the odd pixel goes to the right and to the bottom. An asset whose content
+    /// is one pixel narrower or shorter than the canvas in parity cannot be centred exactly, and
+    /// expecting a half-pixel here would make item 7 unsatisfiable for every such row rather than
+    /// exact for any of them. This is not a tolerance: the comparison is still an exact integer one,
+    /// against the placement `15` §B4 step 2 actually produces.
+    /// </para>
     /// </remarks>
     /// <param name="image">The processed image.</param>
     /// <param name="content">The content's bounding box.</param>
@@ -133,11 +143,11 @@ public sealed class CanvasAndPivotCheck : IQaCheck
         }
 
         // Horizontally centred for both of `15` §C's pivots; only the vertical rule differs.
-        var expectedLeft = (image.Width - content.Width) / 2d;
+        var expectedLeft = (image.Width - content.Width) / 2;
         var expectedTop = pivot switch
         {
-            Doc15Pivots.Center => (image.Height - content.Height) / 2d,
-            Doc15Pivots.BottomCenter => image.Height - (double)content.Height,
+            Doc15Pivots.Center => (image.Height - content.Height) / 2,
+            Doc15Pivots.BottomCenter => image.Height - content.Height,
             _ => throw new InvalidOperationException(
                 $"`15` §C authorises {string.Join(" and ", Doc15Pivots.All)} and nothing else, and " +
                 $"this row declares '{pivot}'. There is no third convention to measure against."),
